+++ packages/iov-ethereum/src/ethereumconnection.spec.ts
@@ -85,7 +85,7 @@ describe("EthereumConnection", () => {
       gasLimit: testConfig.gasLimit,
       memo: `Some text ${Math.random()}`,
     };
-    const signedTransaction = await profile.signTransaction(walletId, sender, sendTx, ethereumCodec, nonce);
+    const signedTransaction = await profile.signTransaction(walletId, sendTx, ethereumCodec, nonce);
     const resultPost = await connection.postTx(ethereumCodec.bytesToPost(signedTransaction));
     return resultPost;
   }
@@ -287,7 +287,7 @@ describe("EthereumConnection", () => {
       };
       const connection = await EthereumConnection.establish(testConfig.base);
       const nonce = await connection.getNonce({ pubkey: mainIdentity.pubkey });
-      const signed = await profile.signTransaction(wallet.id, mainIdentity, sendTx, ethereumCodec, nonce);
+      const signed = await profile.signTransaction(wallet.id, sendTx, ethereumCodec, nonce);
       const bytesToPost = ethereumCodec.bytesToPost(signed);
 
       const result = await connection.postTx(bytesToPost);
@@ -320,7 +320,7 @@ describe("EthereumConnection", () => {
       };
       const connection = await EthereumConnection.establish(testConfig.base);
       const nonce = await connection.getNonce({ pubkey: mainIdentity.pubkey });
-      const signed = await profile.signTransaction(wallet.id, mainIdentity, sendTx, ethereumCodec, nonce);
+      const signed = await profile.signTransaction(wallet.id, sendTx, ethereumCodec, nonce);
       const bytesToPost = ethereumCodec.bytesToPost(signed);
 
       const heightBeforeTransaction = await connection.height();
@@ -366,7 +366,7 @@ describe("EthereumConnection", () => {
         memo: "We \u2665 developers – iov.one",
       };
       const nonce = await connection.getNonce({ pubkey: mainIdentity.pubkey });
-      const signed = await profile.signTransaction(wallet.id, mainIdentity, sendTx, ethereumCodec, nonce);
+      const signed = await profile.signTransaction(wallet.id, sendTx, ethereumCodec, nonce);
       await connection
         .postTx(ethereumCodec.bytesToPost(signed))
         .then(() => fail("must not resolve"))
@@ -398,7 +398,7 @@ describe("EthereumConnection", () => {
         memo: "We \u2665 developers – iov.one",
       };
       const nonce = await connection.getNonce({ pubkey: brokeIdentity.pubkey });
-      const signed = await profile.signTransaction(wallet.id, brokeIdentity, sendTx, ethereumCodec, nonce);
+      const signed = await profile.signTransaction(wallet.id, sendTx, ethereumCodec, nonce);
       await connection
         .postTx(ethereumCodec.bytesToPost(signed))
         .then(() => fail("must not resolve"))
@@ -426,7 +426,7 @@ describe("EthereumConnection", () => {
       };
       const connection = await EthereumConnection.establish(testConfig.base);
       const nonce = await connection.getNonce({ pubkey: mainIdentity.pubkey });
-      const signed = await profile.signTransaction(wallet.id, mainIdentity, sendTx, ethereumCodec, nonce);
+      const signed = await profile.signTransaction(wallet.id, sendTx, ethereumCodec, nonce);
       // tslint:disable-next-line:no-bitwise no-object-mutation
       signed.primarySignature.signature[0] ^= 1;
       // Alternatively we could corrupt the message
@@ -527,7 +527,7 @@ describe("EthereumConnection", () => {
       };
       const connection = await EthereumConnection.establish(testConfig.base);
       const nonce = await connection.getNonce({ pubkey: mainIdentity.pubkey });
-      const signed = await profile.signTransaction(wallet.id, mainIdentity, sendTx, ethereumCodec, nonce);
+      const signed = await profile.signTransaction(wallet.id, sendTx, ethereumCodec, nonce);
       const bytesToPost = ethereumCodec.bytesToPost(signed);
 
       const resultPost = await connection.postTx(bytesToPost);
@@ -601,7 +601,7 @@ describe("EthereumConnection", () => {
         memo: `Search tx test ${new Date()}`,
       };
       const nonce = await connection.getNonce({ pubkey: mainIdentity.pubkey });
-      const signed = await profile.signTransaction(wallet.id, mainIdentity, sendTx, ethereumCodec, nonce);
+      const signed = await profile.signTransaction(wallet.id, sendTx, ethereumCodec, nonce);
       const bytesToPost = ethereumCodec.bytesToPost(signed);
 
       const resultPost = await connection.postTx(bytesToPost);
@@ -786,9 +786,9 @@ describe("EthereumConnection", () => {
 
         const [nonceA, nonceB, nonceC] = await connection.getNonces({ pubkey: sender.pubkey }, 3);
 
-        const signedA = await profile.signTransaction(wallet.id, sender, sendA, ethereumCodec, nonceA);
-        const signedB = await profile.signTransaction(wallet.id, sender, sendB, ethereumCodec, nonceB);
-        const signedC = await profile.signTransaction(wallet.id, sender, sendC, ethereumCodec, nonceC);
+        const signedA = await profile.signTransaction(wallet.id, sendA, ethereumCodec, nonceA);
+        const signedB = await profile.signTransaction(wallet.id, sendB, ethereumCodec, nonceB);
+        const signedC = await profile.signTransaction(wallet.id, sendC, ethereumCodec, nonceC);
         const bytesToPostA = ethereumCodec.bytesToPost(signedA);
         const bytesToPostB = ethereumCodec.bytesToPost(signedB);
         const bytesToPostC = ethereumCodec.bytesToPost(signedC);
@@ -859,9 +859,9 @@ describe("EthereumConnection", () => {
 
         const [nonceA, nonceB, nonceC] = await connection.getNonces({ pubkey: sender.pubkey }, 3);
 
-        const signedA = await profile.signTransaction(wallet.id, sender, sendA, ethereumCodec, nonceA);
-        const signedB = await profile.signTransaction(wallet.id, sender, sendB, ethereumCodec, nonceB);
-        const signedC = await profile.signTransaction(wallet.id, sender, sendC, ethereumCodec, nonceC);
+        const signedA = await profile.signTransaction(wallet.id, sendA, ethereumCodec, nonceA);
+        const signedB = await profile.signTransaction(wallet.id, sendB, ethereumCodec, nonceB);
+        const signedC = await profile.signTransaction(wallet.id, sendC, ethereumCodec, nonceC);
         const bytesToPostA = ethereumCodec.bytesToPost(signedA);
         const bytesToPostB = ethereumCodec.bytesToPost(signedB);
         const bytesToPostC = ethereumCodec.bytesToPost(signedC);
@@ -936,7 +936,7 @@ describe("EthereumConnection", () => {
         };
 
         const nonce = await connection.getNonce({ pubkey: sender.pubkey });
-        const signed = await profile.signTransaction(wallet.id, sender, send, ethereumCodec, nonce);
+        const signed = await profile.signTransaction(wallet.id, send, ethereumCodec, nonce);
         const bytesToPost = ethereumCodec.bytesToPost(signed);
 
         const postResult = await connection.postTx(bytesToPost);
@@ -997,7 +997,7 @@ describe("EthereumConnection", () => {
         };
 
         const nonce = await connection.getNonce({ pubkey: sender.pubkey });
-        const signed = await profile.signTransaction(wallet.id, sender, send, ethereumCodec, nonce);
+        const signed = await profile.signTransaction(wallet.id, send, ethereumCodec, nonce);
         const bytesToPost = ethereumCodec.bytesToPost(signed);
 
         const postResult = await connection.postTx(bytesToPost);