+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Breaking changes
 * @iov/keycontrol: `Keyring.addWallet` now stores a copy of the wallet
 * @iov/keycontrol: Add `.previewIdentity` to the `ReadonlyWallet`/`Wallet` interfaces
 * @iov/keycontrol: Identities now must be unique in `Keyring`.
+* @iov/keycontrol: Remove identity argument from `UserProfile.signTransaction`
 * @iov/ledger-bns: Package removed from this monorepo and now available at
   https://github.com/iov-one/iov-ledger-bns
 * @iov/lisk: Implement `watchBlockHeaders` method on `LiskConnection`.